refactor(551): move site-level Beacon settings out of the platform admin form
The Beacon administration form (gated by the platform-only "administer ys
beacon" permission) mixed platform-operator settings with controls that
site administrators need day to day, leaving site admins unable to reach
them.

Move the content metadata toggle, the guardrail supplement (and its
read-only platform-guardrail display), and the indexing status plus
re-index action into the site-facing Beacon settings form (gated by "manage
ys beacon settings", which site admins hold). The platform admin form now
holds only the sensitive connection and retrieval tuning: the Azure index
connection, sources per answer, relevance threshold, streaming, and the
fallback prompt.

Both forms edit the same ys_beacon.settings object, so config keys and
schema are unchanged; only which screen owns each control moves. No
privilege boundary is weakened: site admins already manage the more
powerful system instructions, and the guardrail supplement can only append
to, never relax, the fixed platform guardrail. The unused entity type
manager dependency is dropped from the admin form.
